moraliser-flask/requirements.txt
# This file is used by pip to install required python packages
# Usage: pip install -r requirements.txt

# Other Packages
python-dotenv

# Flask Framework
Flask==2.2

# Flask Packages
Flask-Login==0.6
Flask-Migrate==3.1
# Flask-Script==2.0.6
Flask-SQLAlchemy==3.0
Flask_RESTful==0.3.9
Flask-Bcrypt==1.0
# Flask-Validator==1.4
Flask-WTF==1.0
# Flask-User==1.0.1.5
flask-cors

# Automated tests
# pytest==3.0.5
# pytest-cov==2.4.0
email_validator

# Datascience Frameworks
numpy
pandas
scikit-learn